requirements.txt
discord.py[voice]== 1.7.3
youtube-dl==2021.6.6
